move tests to a sane place
